Add files via upload
I went through multiple iterations of this project, attempting to display the merge sort algorithm in such a way that small blocks would be sorted from smaller to larger sizes by the algorithm and display their value as an integer in the center of said blocks. I was unable to figure out a method to update the displayed value along with the size of the block due to an issue with aligning the value with a property that may be modified, such as the size in pixels of a block. I am almost certain there is a way to do this through research but it is not something I could find out how to implement and I would be incredibly happy to receive any advice or assistance on the topic. I am by all means pretty new to JS as a whole let alone JSX and HTML and CSS. However through watching videos and doing other online research on the topics covered in this project I feel very happy with my progress and I am definitely looking to continue working with JavaScript and React in the future. Again please feel free to contact me if you are curious at all about this little project or if you have advice to give me on working with this framework. Thank you!